medigenius/requirements.txt
fastapi
langchain
langchain_community
langchain_google_genai
uvicorn
pydantic
pymongo
python-dotenv